requirements.txt
bitsandbytes
dadaptation
diffusers
ipython
lion_pytorch
lpips
matplotlib
numpy
opencv_python
opencv_python_headless
pandas
Pillow
prodigyopt
pydantic
PyYAML
Requests
safetensors
torch
torchvision
tqdm
transformers
xformers
accelerate
peft